requirements.txt
flask
faster-whisper
transformers
torch
spotipy
python-dotenv
ffmpeg-python
